Fix methods and tests
- Change tasks numbering to correct.
- Improve naming of variables and methods.
- In tests add @ParameterizedTest annotations.
- In methods get rid of @SuppressWarnings().
