+++ src/main/java/edu/hw1/Task5.java
@@ -4,13 +4,31 @@ public final class Task5 {
     private Task5() {
     }
 
-    public static String fixString(String input) {
-        char[] characters = input.toCharArray();
-        for (int i = 0; i < characters.length - 1; i += 2) {
-            char temp = characters[i];
-            characters[i] = characters[i + 1];
-            characters[i + 1] = temp;
+    public static boolean isPalindromeDescendant(String number) {
+        String num = number;
+        while (num.length() > 1) {
+            if (isPalindrome(num)) {
+                return true;
+            }
+            num = calculateDescendant(num);
         }
-        return new String(characters);
+        return false;
+    }
+
+    public static boolean isPalindrome(String number) {
+        return number.equals(new StringBuilder(number).reverse().toString());
+    }
+
+    public static String calculateDescendant(String numStr) {
+        StringBuilder result = new StringBuilder();
+        for (int i = 0; i < numStr.length() - 1; i += 2) {
+            int digit1 = Character.getNumericValue(numStr.charAt(i));
+            int digit2 = Character.getNumericValue(numStr.charAt(i + 1));
+            result.append(digit1 + digit2);
+        }
+        if (numStr.length() % 2 == 1) {
+            result.append(numStr.charAt(numStr.length() - 1));
+        }
+        return result.toString();
     }
 }
