+++ src/main/java/edu/hw1/Task6.java
@@ -1,34 +1,71 @@
 package edu.hw1;
 
+import java.util.Arrays;
+
 public final class Task6 {
     private Task6() {
     }
 
-    public static boolean isPalindromeDescendant(String number) {
-        String num = number;
-        while (num.length() > 1) {
-            if (isPalindrome(num)) {
-                return true;
+    private static final int MIN_FOUR_DIGIT_NUMBER = 1000;
+    private static final int MAX_FOUR_DIGIT_NUMBER = 9999;
+    private static final int KAPREKAR_CONSTANT = 6174;
+    private static final int FOUR = 4;
+    private static final int BASE_TEN = 10;
+
+    private static boolean isValidInput(int number) {
+        String numberStr = String.valueOf(number);
+
+        return number > MIN_FOUR_DIGIT_NUMBER && number <= MAX_FOUR_DIGIT_NUMBER
+            && numberStr.chars().distinct().count() != 1;
+    }
+
+    public static int calculateKaprekarConstant(int number) {
+        int value = number;
+        if (!isValidInput(number)) {
+            return -1;
+        }
+
+        if (value == KAPREKAR_CONSTANT) {
+            return 0;
+        } else {
+            int[] digits = new int[FOUR];
+            for (int i = FOUR - 1; i >= 0; i--) {
+                digits[i] = value % BASE_TEN;
+                value /= BASE_TEN;
             }
-            num = calculateDescendant(num);
+            // сортируем цифры в порядке возрастания и убывания
+            int ascending = sortDigits(digits, true);
+            int descending = sortDigits(digits, false);
+            int diff = descending - ascending;
+            // ррекурсивный вызов
+            int descendant = calculateKaprekarConstant(diff);
+            if (descendant == -1) {
+                return -1;
+            } else {
+                return 1 + descendant;
+            }
         }
-        return false;
     }
 
-    public static boolean isPalindrome(String number) {
-        return number.equals(new StringBuilder(number).reverse().toString());
-    }
-
-    public static String calculateDescendant(String numStr) {
-        StringBuilder result = new StringBuilder();
-        for (int i = 0; i < numStr.length() - 1; i += 2) {
-            int digit1 = Character.getNumericValue(numStr.charAt(i));
-            int digit2 = Character.getNumericValue(numStr.charAt(i + 1));
-            result.append(digit1 + digit2);
+    public static int sortDigits(int[] digits, boolean ascending) {
+        if (ascending) {
+            Arrays.sort(digits);
+        } else {
+            Arrays.sort(digits);
+            int i = 0;
+            int j = FOUR - 1;
+            while (i < j) {
+                digits[i] += digits[j];
+                digits[j] = digits[i] - digits[j];
+                digits[i] -= digits[j];
+                i++;
+                j--;
+            }
         }
-        if (numStr.length() % 2 == 1) {
-            result.append(numStr.charAt(numStr.length() - 1));
+        int number = 0;
+        for (int digit : digits) {
+            number = number * BASE_TEN + digit;
         }
-        return result.toString();
+        return number;
     }
 }
