+++ src/test/java/edu/hw1/Task2Test.java
@@ -1,34 +1,14 @@
 package edu.hw1;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.params.provider.CsvSource;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class Task2Test {
 
-    @Test
-    public void testValidInput() {
-        assertEquals(60, Task2.minutesToSeconds("01:00"));
-        assertEquals(836, Task2.minutesToSeconds("13:56"));
-        assertEquals(59999, Task2.minutesToSeconds("999:59"));
-    }
-
     @ParameterizedTest
-    @ValueSource(strings = {"10:60", "20:70"})
-    public void testInvalidSeconds(String input) {
-        assertEquals(-1, Task2.minutesToSeconds(input));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"123", "12:34:56", "12:", ":34", "12:34:"})
-    public void testInvalidFormat(String input) {
-        assertEquals(-1, Task2.minutesToSeconds(input));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"12:34a", "12:3b"})
-    public void testInvalidCharacters(String input) {
-        assertEquals(-1, Task2.minutesToSeconds(input));
+    @CsvSource({"4666, 4", "544, 3", "0, 1", "-123, 3"})
+    public void testCountDigits(int input, int expected) {
+        assertThat(Task2.countDigits(input)).isEqualTo(expected);
     }
 }
